[Cleanup] Split settings from global.h into separate .h and .cpp
