+++ src/DataStructs/SecurityStruct.h
@@ -1,0 +1,53 @@
+#ifndef DATASTRUCTS_SECURITYSTRUCT_H
+#define DATASTRUCTS_SECURITYSTRUCT_H
+
+
+/*********************************************************************************************\
+ * SecurityStruct
+\*********************************************************************************************/
+struct SecurityStruct
+{
+  SecurityStruct() {
+    ZERO_FILL(WifiSSID);
+    ZERO_FILL(WifiKey);
+    ZERO_FILL(WifiSSID2);
+    ZERO_FILL(WifiKey2);
+    ZERO_FILL(WifiAPKey);
+    for (byte i = 0; i < CONTROLLER_MAX; ++i) {
+      ZERO_FILL(ControllerUser[i]);
+      ZERO_FILL(ControllerPassword[i]);
+    }
+    ZERO_FILL(Password);
+  }
+
+  void validate() {
+    ZERO_TERMINATE(WifiSSID);
+    ZERO_TERMINATE(WifiKey);
+    ZERO_TERMINATE(WifiSSID2);
+    ZERO_TERMINATE(WifiKey2);
+    ZERO_TERMINATE(WifiAPKey);
+    for (byte i = 0; i < CONTROLLER_MAX; ++i) {
+      ZERO_TERMINATE(ControllerUser[i]);
+      ZERO_TERMINATE(ControllerPassword[i]);
+    }
+    ZERO_TERMINATE(Password);
+  }
+
+  char          WifiSSID[32];
+  char          WifiKey[64];
+  char          WifiSSID2[32];
+  char          WifiKey2[64];
+  char          WifiAPKey[64];
+  char          ControllerUser[CONTROLLER_MAX][26];
+  char          ControllerPassword[CONTROLLER_MAX][64];
+  char          Password[26];
+  byte          AllowedIPrangeLow[4]; // TD-er: Use these
+  byte          AllowedIPrangeHigh[4];
+  byte          IPblockLevel;
+
+  //its safe to extend this struct, up to 4096 bytes, default values in config are 0. Make sure crc is last
+  uint8_t       ProgmemMd5[16] = {0}; // crc of the binary that last saved the struct to file.
+  uint8_t       md5[16] = {0};
+};
+
+#endif // DATASTRUCTS_SECURITYSTRUCT_H
