+++ src/ESPEasy-Globals.h
@@ -10,14 +10,10 @@
 #pragma GCC system_header
 #endif
 
-#include <stddef.h>
+#include "ESPEasy_common.h"
+#include "ESPEasy_fdwdecl.h"
 
-namespace std
-{
-  using ::ptrdiff_t;
-  using ::size_t;
-}
-
+#include "DataStructs/ESPEasyLimits.h"
 
 #include <FS.h>
 
@@ -37,124 +33,9 @@ void check_size() {
 }
 
 
+// User configuration
+#include "DataStructs/ESPEasyDefaults.h"
 
-// ********************************************************************************
-//   User specific configuration
-// ********************************************************************************
-
-// Set default configuration settings if you want (not mandatory)
-// You can always change these during runtime and save to eeprom
-// After loading firmware, issue a 'reset' command to load the defaults.
-// --- Basic Config Settings ------------------------------------------------------------------------
-#define DEFAULT_NAME        "ESP_Easy"          // Enter your device friendly name
-#define UNIT                            0                                       // Unit Number
-#define DEFAULT_DELAY       60                  // Sleep Delay in seconds
-
-// --- Wifi AP Mode (when your Wifi Network is not reachable) ----------------------------------------
-#define DEFAULT_AP_IP       192,168,4,1         // Enter IP address (comma separated) for AP (config) mode
-#define DEFAULT_AP_SUBNET   255,255,255,0       // Enter IP address (comma separated) for AP (config) mode
-#define DEFAULT_AP_KEY      "configesp"         // Enter network WPA key for AP (config) mode
-
-// --- Wifi Client Mode -----------------------------------------------------------------------------
-#define DEFAULT_SSID        "ssid"              // Enter your Wifi network SSID
-#define DEFAULT_KEY         "wpakey"            // Enter your Wifi network WPA key
-
-#define DEFAULT_USE_STATIC_IP   false           // (true|false) enabled or disabled static IP
-#define DEFAULT_IP          "192.168.0.50"      // Enter your IP address
-#define DEFAULT_DNS         "192.168.0.1"       // Enter your DNS
-#define DEFAULT_GW          "192.168.0.1"       // Enter your Gateway
-#define DEFAULT_SUBNET      "255.255.255.0"     // Enter your Subnet
-#define DEFAULT_IPRANGE_LOW  "0.0.0.0"          // Allowed IP range to access webserver
-#define DEFAULT_IPRANGE_HIGH "255.255.255.255"  // Allowed IP range to access webserver
-#define DEFAULT_IP_BLOCK_LEVEL 1                // 0: ALL_ALLOWED  1: LOCAL_SUBNET_ALLOWED  2: ONLY_IP_RANGE_ALLOWED
-#define DEFAULT_ADMIN_USERNAME  "admin"
-
-#define DEFAULT_WIFI_CONNECTION_TIMEOUT  10000  // minimum timeout in ms for WiFi to be connected.
-#define DEFAULT_WIFI_FORCE_BG_MODE       false  // when set, only allow to connect in 802.11B or G mode (not N)
-#define DEFAULT_WIFI_RESTART_WIFI_CONN_LOST  false // Perform wifi off and on when connection was lost.
-#define DEFAULT_ECO_MODE                 false   // When set, make idle calls between executing tasks.
-#define DEFAULT_WIFI_NONE_SLEEP          false  // When set, the wifi will be set to no longer sleep (more power used and need reboot to reset mode)
-#define DEFAULT_GRATUITOUS_ARD           false  // When set, the node will send periodical gratuitous ARP packets to announce itself.
-
-// --- Default Controller ------------------------------------------------------------------------------
-#define DEFAULT_CONTROLLER   false              // true or false enabled or disabled, set 1st controller defaults
-// using a default template, you also need to set a DEFAULT PROTOCOL to a suitable MQTT protocol !
-#define DEFAULT_PUB         "sensors/espeasy/%sysname%/%tskname%/%valname%" // Enter your pub
-#define DEFAULT_SUB         "sensors/espeasy/%sysname%/#" // Enter your sub
-#define DEFAULT_SERVER      "192.168.0.8"       // Enter your Server IP address
-#define DEFAULT_PORT        8080                // Enter your Server port value
-
-#define DEFAULT_PROTOCOL    0                   // Protocol used for controller communications
-//   0 = Stand-alone (no controller set)
-//   1 = Domoticz HTTP
-//   2 = Domoticz MQTT
-//   3 = Nodo Telnet
-//   4 = ThingSpeak
-//   5 = Home Assistant (openHAB) MQTT
-//   6 = PiDome MQTT
-//   7 = EmonCMS
-//   8 = Generic HTTP
-//   9 = FHEM HTTP
-
-#define DEFAULT_PIN_I2C_SDA              4
-#define DEFAULT_PIN_I2C_SCL              5
-
-#define DEFAULT_PIN_STATUS_LED           -1
-#define DEFAULT_PIN_STATUS_LED_INVERSED  true
-
-
-
-// --- Advanced Settings ---------------------------------------------------------------------------------
-#if defined(ESP32)
-  #define USE_RTOS_MULTITASKING
-#endif
-#ifdef M5STACK_ESP
-//  #include <M5Stack.h>
-#endif
-
-#define DEFAULT_USE_RULES                       false   // (true|false) Enable Rules?
-#define DEFAULT_RULES_OLDENGINE                true
-
-#define DEFAULT_MQTT_RETAIN                     false   // (true|false) Retain MQTT messages?
-#define DEFAULT_MQTT_DELAY                      100    // Time in milliseconds to retain MQTT messages
-#define DEFAULT_MQTT_LWT_TOPIC                  ""      // Default lwt topic
-#define DEFAULT_MQTT_LWT_CONNECT_MESSAGE        "Connected" // Default lwt message
-#define DEFAULT_MQTT_LWT_DISCONNECT_MESSAGE     "Connection Lost" // Default lwt message
-#define DEFAULT_MQTT_USE_UNITNAME_AS_CLIENTID   0
-
-#define DEFAULT_USE_NTP                         false   // (true|false) Use NTP Server
-#define DEFAULT_NTP_HOST                        ""              // NTP Server Hostname
-#define DEFAULT_TIME_ZONE                       0               // Time Offset (in minutes)
-#define DEFAULT_USE_DST                         false   // (true|false) Use Daily Time Saving
-
-#define LOG_TO_SERIAL         1
-#define LOG_TO_SYSLOG         2
-#define LOG_TO_WEBLOG         3
-#define LOG_TO_SDCARD         4
-#define DEFAULT_SYSLOG_IP                       ""                      // Syslog IP Address
-#define DEFAULT_SYSLOG_LEVEL            0                               // Syslog Log Level
-#define DEFAULT_SERIAL_LOG_LEVEL        LOG_LEVEL_INFO                  // Serial Log Level
-#define DEFAULT_WEB_LOG_LEVEL           LOG_LEVEL_INFO                  // Web Log Level
-#define DEFAULT_SD_LOG_LEVEL            0                               // SD Card Log Level
-#define DEFAULT_USE_SD_LOG                      false                   // (true|false) Enable Logging to the SD card
-
-#define DEFAULT_USE_SERIAL                      true    // (true|false) Enable Logging to the Serial Port
-#define DEFAULT_SERIAL_BAUD                     115200  // Serial Port Baud Rate
-
-#define DEFAULT_SYSLOG_FACILITY 	0 	// kern
-
-/*
-// --- Experimental Advanced Settings (NOT ACTIVES at this time) ------------------------------------
-
-#define DEFAULT_USE_GLOBAL_SYNC                 false           // (true|false)
-#define DEFAULT_SYNC_UDP_PORT                   0                       //
-
-#define DEFAULT_IP_OCTET                                0                       //
-#define DEFAULT_WD_IC2_ADDRESS                  0                       //
-#define DEFAULT_USE_SSDP                        false           // (true|false)
-#define DEFAULT_CON_FAIL_THRES                  0                       //
-#define DEFAULT_I2C_CLOCK_LIMIT                 0                       //
-*/
 
 
 // Enable FEATURE_ADC_VCC to measure supply voltage using the analog pin
@@ -272,19 +153,6 @@ void check_size() {
 #define TIMING_STATS_THRESHOLD             100000
 #define TIMER_GRATUITOUS_ARP_MAX           5000
 
-// Minimum delay between messages for a controller to send in msec.
-#define CONTROLLER_DELAY_QUEUE_DELAY_MAX   3600000
-#define CONTROLLER_DELAY_QUEUE_DELAY_DFLT  100
-// Queue length for controller messages not yet sent.
-#define CONTROLLER_DELAY_QUEUE_DEPTH_MAX   50
-#define CONTROLLER_DELAY_QUEUE_DEPTH_DFLT  10
-// Number of retries to send a message by a controller.
-// N.B. Retries without a connection to wifi do not count as retry.
-#define CONTROLLER_DELAY_QUEUE_RETRY_MAX   10
-#define CONTROLLER_DELAY_QUEUE_RETRY_DFLT  10
-// Timeout of the client in msec.
-#define CONTROLLER_CLIENTTIMEOUT_MAX     1000
-#define CONTROLLER_CLIENTTIMEOUT_DFLT     100
 
 
 #define PLUGIN_INIT_ALL                     1
@@ -382,47 +250,7 @@ void check_size() {
 #define CMD_REBOOT                         89
 #define CMD_WIFI_DISCONNECT               135
 
-#if defined(PLUGIN_BUILD_TESTING) || defined(PLUGIN_BUILD_DEV)
-  #define DEVICES_MAX                      95
-#else
-  #ifdef ESP32
-    #define DEVICES_MAX                      85
-  #else
-    #define DEVICES_MAX                      60
-  #endif
-#endif
 
-#if defined(ESP8266)
-  #define TASKS_MAX                          12 // max 12!
-  #define MAX_GPIO                           16
-#endif
-#if defined(ESP32)
-  #define TASKS_MAX                          32
-  #define MAX_GPIO                           39
-#endif
-
-#define CONTROLLER_MAX                      3 // max 4!
-#define NOTIFICATION_MAX                    3 // max 4!
-#define VARS_PER_TASK                       4
-#define PLUGIN_MAX                DEVICES_MAX
-#define PLUGIN_CONFIGVAR_MAX                8
-#define PLUGIN_CONFIGFLOATVAR_MAX           4
-#define PLUGIN_CONFIGLONGVAR_MAX            4
-#define PLUGIN_EXTRACONFIGVAR_MAX          16
-#define CPLUGIN_MAX                        20
-#define NPLUGIN_MAX                         4
-#define UNIT_MAX                          254 // unit 255 = broadcast
-#define RULES_TIMER_MAX                     8
-//#define PINSTATE_TABLE_MAX                 32
-#define RULES_MAX_SIZE                   2048
-#define RULES_MAX_NESTING_LEVEL             3
-#define RULESETS_MAX                        4
-#define RULES_BUFFER_SIZE                  64
-#define NAME_FORMULA_LENGTH_MAX            40
-#define RULES_IF_MAX_NESTING_LEVEL          4
-#define CUSTOM_VARS_MAX                    16
-
-#define UDP_PACKETSIZE_MAX               2048
 
 #define PIN_MODE_UNDEFINED                  0
 #define PIN_MODE_INPUT                      1
@@ -495,25 +323,6 @@ void check_size() {
   #define CONFIG_FILE_SIZE               131072
 #endif
 
-#define ZERO_FILL(S)  memset((S), 0, sizeof(S))
-#define ZERO_TERMINATE(S)  S[sizeof(S) - 1] = 0
-
-enum SettingsType {
-  BasicSettings_Type = 0,
-  TaskSettings_Type,
-  CustomTaskSettings_Type,
-  ControllerSettings_Type,
-  CustomControllerSettings_Type,
-  NotificationSettings_Type,
-
-  SettingsType_MAX
-
-};
-String getSettingsTypeString(SettingsType settingsType);
-bool getSettingsParameters(SettingsType settingsType, int index, int& offset, int& max_size);
-#ifndef BUILD_MINIMAL_OTA
-bool showSettingsFileLayout = false;
-#endif
 
 /*
         To modify the stock configuration without changing this repo file :
@@ -524,6 +333,22 @@ bool showSettingsFileLayout = false;
 #ifdef USE_CUSTOM_H
 #include "Custom.h"
 #endif
+
+#include "DataStructs/ESPEasyLimits.h"
+
+#include "DataStructs/ESPEasy_EventStruct.h"
+#include "DataStructs/SettingsType.h"
+#include "DataStructs/CRCStruct.h"
+#include "DataStructs/SecurityStruct.h"
+#include "DataStructs/SettingsStruct.h"
+#include "DataStructs/DeviceModel.h"
+#include "DataStructs/FactoryDefaultPref.h"
+#include "DataStructs/ControllerSettingsStruct.h"
+
+CRCStruct CRCValues;
+SecurityStruct SecuritySettings;
+SettingsStruct Settings;
+ResetFactoryDefaultPreference_struct ResetFactoryDefaultPreference;
 
 #include "define_plugin_sets.h"
 #include "WebStaticData.h"
@@ -635,9 +460,6 @@ bool showSettingsFileLayout = false;
 #include <DNSServer.h>
 #include <Wire.h>
 #include <SPI.h>
-#ifdef USES_MQTT
-#include <PubSubClient.h>
-#endif //USES_MQTT
 #include <FS.h>
 #ifdef FEATURE_SD
 #include <SD.h>
@@ -647,6 +469,24 @@ using namespace fs;
 #include <base64.h>
 #if FEATURE_ADC_VCC
 ADC_MODE(ADC_VCC);
+#endif
+
+
+I2Cdev i2cdev;
+
+#ifdef USES_MQTT
+#include <PubSubClient.h>
+// MQTT client
+WiFiClient mqtt;
+PubSubClient MQTTclient(mqtt);
+bool MQTTclient_should_reconnect = true;
+bool MQTTclient_connected = false;
+int mqtt_reconnect_count = 0;
+#endif //USES_MQTT
+
+#ifdef USES_P037
+// mqtt import status
+bool P037_MQTTImport_connected = false;
 #endif
 
 #define ESPEASY_WIFI_DISCONNECTED            0
@@ -671,89 +511,14 @@ IPAddress apIP(DEFAULT_AP_IP);
 DNSServer dnsServer;
 bool dnsServerActive = false;
 
-#ifdef USES_MQTT
-// MQTT client
-WiFiClient mqtt;
-PubSubClient MQTTclient(mqtt);
-bool MQTTclient_should_reconnect = true;
-bool MQTTclient_connected = false;
-int mqtt_reconnect_count = 0;
-#endif //USES_MQTT
-
 //NTP status
 bool statusNTPInitialized = false;
-
-// mqtt import status
-bool P037_MQTTImport_connected = false;
 
 // udp protocol stuff (syslog, global sync, node info list, ntp time)
 WiFiUDP portUDP;
 
 class TimingStats;
 
-/*********************************************************************************************\
- * CRCStruct
-\*********************************************************************************************/
-struct CRCStruct{
-  char compileTimeMD5[16+32+1]= "MD5_MD5_MD5_MD5_BoundariesOfTheSegmentsGoHere...";
-  char binaryFilename[32+32+1]= "ThisIsTheDummyPlaceHolderForTheBinaryFilename64ByteLongFilenames";
-  char compileTime[16]= __TIME__;
-  char compileDate[16]= __DATE__;
-  uint8_t runTimeMD5[16]={0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
-  bool checkPassed (void){ return memcmp(compileTimeMD5,runTimeMD5,16)==0 ; }
-  uint32_t numberOfCRCBytes=0;
-}CRCValues;
-
-
-
-
-/*********************************************************************************************\
- * SecurityStruct
-\*********************************************************************************************/
-struct SecurityStruct
-{
-  SecurityStruct() {
-    ZERO_FILL(WifiSSID);
-    ZERO_FILL(WifiKey);
-    ZERO_FILL(WifiSSID2);
-    ZERO_FILL(WifiKey2);
-    ZERO_FILL(WifiAPKey);
-    for (byte i = 0; i < CONTROLLER_MAX; ++i) {
-      ZERO_FILL(ControllerUser[i]);
-      ZERO_FILL(ControllerPassword[i]);
-    }
-    ZERO_FILL(Password);
-  }
-
-  void validate() {
-    ZERO_TERMINATE(WifiSSID);
-    ZERO_TERMINATE(WifiKey);
-    ZERO_TERMINATE(WifiSSID2);
-    ZERO_TERMINATE(WifiKey2);
-    ZERO_TERMINATE(WifiAPKey);
-    for (byte i = 0; i < CONTROLLER_MAX; ++i) {
-      ZERO_TERMINATE(ControllerUser[i]);
-      ZERO_TERMINATE(ControllerPassword[i]);
-    }
-    ZERO_TERMINATE(Password);
-  }
-
-  char          WifiSSID[32];
-  char          WifiKey[64];
-  char          WifiSSID2[32];
-  char          WifiKey2[64];
-  char          WifiAPKey[64];
-  char          ControllerUser[CONTROLLER_MAX][26];
-  char          ControllerPassword[CONTROLLER_MAX][64];
-  char          Password[26];
-  byte          AllowedIPrangeLow[4]; // TD-er: Use these
-  byte          AllowedIPrangeHigh[4];
-  byte          IPblockLevel;
-
-  //its safe to extend this struct, up to 4096 bytes, default values in config are 0. Make sure crc is last
-  uint8_t       ProgmemMd5[16] = {0}; // crc of the binary that last saved the struct to file.
-  uint8_t       md5[16] = {0};
-} SecuritySettings;
 
 
 
@@ -767,460 +532,9 @@ struct SecurityStruct
 \*********************************************************************************************/
 float         customFloatVar[CUSTOM_VARS_MAX];
 
-/*********************************************************************************************\
- * SettingsStruct
-\*********************************************************************************************/
-struct SettingsStruct
-{
-  SettingsStruct() : ResetFactoryDefaultPreference(0) {
-    clearAll();
-    clearNetworkSettings();
-  }
-
-  // VariousBits1 defaults to 0, keep in mind when adding bit lookups.
-  bool appendUnitToHostname() {  return !getBitFromUL(VariousBits1, 1); }
-  void appendUnitToHostname(bool value) { setBitToUL(VariousBits1, 1, !value); }
-
-  bool uniqueMQTTclientIdReconnect() {  return getBitFromUL(VariousBits1, 2); }
-  void uniqueMQTTclientIdReconnect(bool value) { setBitToUL(VariousBits1, 2, value); }
-
-  bool OldRulesEngine() {  return !getBitFromUL(VariousBits1, 3); }
-  void OldRulesEngine(bool value) {  setBitToUL(VariousBits1, 3, !value); }
-
-  bool ForceWiFi_bg_mode() {  return getBitFromUL(VariousBits1, 4); }
-  void ForceWiFi_bg_mode(bool value) {  setBitToUL(VariousBits1, 4, value); }
-
-  bool WiFiRestart_connection_lost() {  return getBitFromUL(VariousBits1, 5); }
-  void WiFiRestart_connection_lost(bool value) {  setBitToUL(VariousBits1, 5, value); }
-
-  bool EcoPowerMode() {  return getBitFromUL(VariousBits1, 6); }
-  void EcoPowerMode(bool value) {  setBitToUL(VariousBits1, 6, value); }
-
-  bool WifiNoneSleep() {  return getBitFromUL(VariousBits1, 7); }
-  void WifiNoneSleep(bool value) {  setBitToUL(VariousBits1, 7, value); }
-
-  // Enable send gratuitous ARP by default, so invert the values (default = 0)
-  bool gratuitousARP() {  return !getBitFromUL(VariousBits1, 8); }
-  void gratuitousARP(bool value) {  setBitToUL(VariousBits1, 8, !value); }
-
-  void validate() {
-    if (UDPPort > 65535) UDPPort = 0;
-
-    if (Latitude  < -90.0  || Latitude > 90.0) Latitude = 0.0;
-    if (Longitude < -180.0 || Longitude > 180.0) Longitude = 0.0;
-    if (VariousBits1 > (1 << 30)) VariousBits1 = 0;
-    ZERO_TERMINATE(Name);
-    ZERO_TERMINATE(NTPHost);
-  }
-
-  bool networkSettingsEmpty() {
-    return (IP[0] == 0 && Gateway[0] == 0 && Subnet[0] == 0 && DNS[0] == 0);
-  }
-
-  void clearNetworkSettings() {
-    for (byte i = 0; i < 4; ++i) {
-      IP[i] = 0;
-      Gateway[i] = 0;
-      Subnet[i] = 0;
-      DNS[i] = 0;
-    }
-  }
-
-  void clearTimeSettings() {
-    UseNTP = false;
-    ZERO_FILL(NTPHost);
-    TimeZone = 0;
-    DST = false;
-    DST_Start = 0;
-    DST_End = 0;
-    Latitude = 0.0;
-    Longitude = 0.0;
-  }
-
-  void clearNotifications() {
-    for (byte i = 0; i < NOTIFICATION_MAX; ++i) {
-      Notification[i] = 0;
-      NotificationEnabled[i] = false;
-    }
-  }
-
-  void clearControllers() {
-    for (byte i = 0; i < CONTROLLER_MAX; ++i) {
-      Protocol[i] = 0;
-      ControllerEnabled[i] = false;
-    }
-  }
-
-  void clearTasks() {
-    for (byte task = 0; task < TASKS_MAX; ++task) {
-      clearTask(task);
-    }
-  }
-
-  void clearLogSettings() {
-    SyslogLevel = 0;
-    SerialLogLevel = 0;
-    WebLogLevel = 0;
-    SDLogLevel = 0;
-    SyslogFacility = DEFAULT_SYSLOG_FACILITY;
-    for (byte i = 0; i < 4; ++i) {  Syslog_IP[i] = 0; }
-  }
-
-  void clearUnitNameSettings() {
-    Unit = 0;
-    ZERO_FILL(Name);
-    UDPPort = 0;
-  }
-
-  void clearMisc() {
-    PID = 0;
-    Version = 0;
-    Build = 0;
-    IP_Octet = 0;
-    Delay = 0;
-    Pin_i2c_sda = -1;
-    Pin_i2c_scl = -1;
-    Pin_status_led = -1;
-    Pin_sd_cs = -1;
-    for (byte i = 0; i < 17; ++i) { PinBootStates[i] = 0; }
-    BaudRate = 0;
-    MessageDelay = 0;
-    deepSleep = 0;
-    CustomCSS = false;
-    WDI2CAddress = 0;
-    UseRules = false;
-    UseSerial = true;
-    UseSSDP = false;
-    WireClockStretchLimit = 0;
-    GlobalSync = false;
-    ConnectionFailuresThreshold = 0;
-    MQTTRetainFlag = false;
-    InitSPI = false;
-    Pin_status_led_Inversed = false;
-    deepSleepOnFail = false;
-    UseValueLogger = false;
-    ArduinoOTAEnable = false;
-    UseRTOSMultitasking = false;
-    Pin_Reset = -1;
-    StructSize = sizeof(SettingsStruct);
-    MQTTUseUnitNameAsClientId = 0;
-    VariousBits1 = 0;
-    OldRulesEngine(DEFAULT_RULES_OLDENGINE);
-    ForceWiFi_bg_mode(DEFAULT_WIFI_FORCE_BG_MODE);
-    WiFiRestart_connection_lost(DEFAULT_WIFI_RESTART_WIFI_CONN_LOST);
-    EcoPowerMode(DEFAULT_ECO_MODE);
-    WifiNoneSleep(DEFAULT_WIFI_NONE_SLEEP);
-    gratuitousARP(DEFAULT_GRATUITOUS_ARD);
-  }
-
-  void clearAll() {
-    clearMisc();
-    clearTimeSettings();
-    clearNetworkSettings();
-    clearNotifications();
-    clearControllers();
-    clearTasks();
-    clearLogSettings();
-    clearUnitNameSettings();
-  }
-
-  void clearTask(byte task) {
-    for (byte i = 0; i < CONTROLLER_MAX; ++i) {
-      TaskDeviceID[i][task] = 0;
-      TaskDeviceSendData[i][task] = false;
-    }
-    TaskDeviceNumber[task] = 0;
-    OLD_TaskDeviceID[task] = 0; //UNUSED: this can be removed
-    TaskDevicePin1[task] = -1;
-    TaskDevicePin2[task] = -1;
-    TaskDevicePin3[task] = -1;
-    TaskDevicePort[task] = 0;
-    TaskDevicePin1PullUp[task] = false;
-    for (byte cv = 0; cv < PLUGIN_CONFIGVAR_MAX; ++cv) {
-      TaskDevicePluginConfig[task][cv] = 0;
-    }
-    TaskDevicePin1Inversed[task] = false;
-    for (byte cv = 0; cv < PLUGIN_CONFIGFLOATVAR_MAX; ++cv) {
-      TaskDevicePluginConfigFloat[task][cv] = 0.0;
-    }
-    for (byte cv = 0; cv < PLUGIN_CONFIGLONGVAR_MAX; ++cv) {
-      TaskDevicePluginConfigLong[task][cv] = 0;
-    }
-    OLD_TaskDeviceSendData[task] = false;
-    TaskDeviceGlobalSync[task] = false;
-    TaskDeviceDataFeed[task] = 0;
-    TaskDeviceTimer[task] = 0;
-    TaskDeviceEnabled[task] = false;
-  }
-
-  unsigned long PID;
-  int           Version;
-  int16_t       Build;
-  byte          IP[4];
-  byte          Gateway[4];
-  byte          Subnet[4];
-  byte          DNS[4];
-  byte          IP_Octet;
-  byte          Unit;
-  char          Name[26];
-  char          NTPHost[64];
-  unsigned long Delay;              // Sleep time in seconds
-  int8_t        Pin_i2c_sda;
-  int8_t        Pin_i2c_scl;
-  int8_t        Pin_status_led;
-  int8_t        Pin_sd_cs;
-  int8_t        PinBootStates[17];  // FIXME TD-er: this is ESP8266 number of pins. ESP32 has double.
-  byte          Syslog_IP[4];
-  unsigned int  UDPPort;
-  byte          SyslogLevel;
-  byte          SerialLogLevel;
-  byte          WebLogLevel;
-  byte          SDLogLevel;
-  unsigned long BaudRate;
-  unsigned long MessageDelay;
-  byte          deepSleep;   // 0 = Sleep Disabled, else time awake from sleep in seconds
-  boolean       CustomCSS;
-  boolean       DST;
-  byte          WDI2CAddress;
-  boolean       UseRules;
-  boolean       UseSerial;
-  boolean       UseSSDP;
-  boolean       UseNTP;
-  unsigned long WireClockStretchLimit;
-  boolean       GlobalSync;
-  unsigned long ConnectionFailuresThreshold;
-  int16_t       TimeZone;
-  boolean       MQTTRetainFlag;
-  boolean       InitSPI;
-  byte          Protocol[CONTROLLER_MAX];
-  byte          Notification[NOTIFICATION_MAX]; //notifications, point to a NPLUGIN id
-  byte          TaskDeviceNumber[TASKS_MAX]; // The "plugin number" set at as task (e.g. 4 for P004_dallas)
-  unsigned int  OLD_TaskDeviceID[TASKS_MAX];  //UNUSED: this can be removed
-  union {
-    struct {
-      int8_t        TaskDevicePin1[TASKS_MAX];
-      int8_t        TaskDevicePin2[TASKS_MAX];
-      int8_t        TaskDevicePin3[TASKS_MAX];
-      byte          TaskDevicePort[TASKS_MAX];
-    };
-    int8_t        TaskDevicePin[4][TASKS_MAX];
-  };
-  boolean       TaskDevicePin1PullUp[TASKS_MAX];
-  int16_t       TaskDevicePluginConfig[TASKS_MAX][PLUGIN_CONFIGVAR_MAX];
-  boolean       TaskDevicePin1Inversed[TASKS_MAX];
-  float         TaskDevicePluginConfigFloat[TASKS_MAX][PLUGIN_CONFIGFLOATVAR_MAX];
-  long          TaskDevicePluginConfigLong[TASKS_MAX][PLUGIN_CONFIGLONGVAR_MAX];
-  boolean       OLD_TaskDeviceSendData[TASKS_MAX];
-  boolean       TaskDeviceGlobalSync[TASKS_MAX];
-  byte          TaskDeviceDataFeed[TASKS_MAX];    // When set to 0, only read local connected sensorsfeeds
-  unsigned long TaskDeviceTimer[TASKS_MAX];
-  boolean       TaskDeviceEnabled[TASKS_MAX];
-  boolean       ControllerEnabled[CONTROLLER_MAX];
-  boolean       NotificationEnabled[NOTIFICATION_MAX];
-  unsigned int  TaskDeviceID[CONTROLLER_MAX][TASKS_MAX];        // IDX number (mainly used by Domoticz)
-  boolean       TaskDeviceSendData[CONTROLLER_MAX][TASKS_MAX];
-  boolean       Pin_status_led_Inversed;
-  boolean       deepSleepOnFail;
-  boolean       UseValueLogger;
-  boolean       ArduinoOTAEnable;
-  uint16_t      DST_Start;
-  uint16_t      DST_End;
-  boolean       UseRTOSMultitasking;
-  int8_t        Pin_Reset;
-  byte          SyslogFacility;
-  uint32_t      StructSize;  // Forced to be 32 bit, to make sure alignment is clear.
-  boolean       MQTTUseUnitNameAsClientId;
-
-  //its safe to extend this struct, up to several bytes, default values in config are 0
-  //look in misc.ino how config.dat is used because also other stuff is stored in it at different offsets.
-  //TODO: document config.dat somewhere here
-  float         Latitude;
-  float         Longitude;
-  uint32_t      VariousBits1;
-  uint32_t      ResetFactoryDefaultPreference; // Do not clear this one in the clearAll()
-
-  // FIXME @TD-er: As discussed in #1292, the CRC for the settings is now disabled.
-  // make sure crc is the last value in the struct
-  // Try to extend settings to make the checksum 4-byte aligned.
-//  uint8_t       ProgmemMd5[16]; // crc of the binary that last saved the struct to file.
-//  uint8_t       md5[16];
-} Settings;
-/*
-SettingsStruct* SettingsStruct_ptr = new SettingsStruct;
-SettingsStruct& Settings = *SettingsStruct_ptr;
-*/
 
 
-/*********************************************************************************************\
- * ControllerSettingsStruct definition
-\*********************************************************************************************/
-struct ControllerSettingsStruct
-{
-  ControllerSettingsStruct()
-  {
-    reset();
-  }
 
-  void reset() {
-    UseDNS = false;
-    Port = 0;
-    MinimalTimeBetweenMessages = CONTROLLER_DELAY_QUEUE_DELAY_DFLT;
-    MaxQueueDepth = CONTROLLER_DELAY_QUEUE_DEPTH_DFLT;
-    MaxRetry = CONTROLLER_DELAY_QUEUE_RETRY_DFLT;
-    DeleteOldest = false;
-    ClientTimeout = CONTROLLER_CLIENTTIMEOUT_DFLT;
-    MustCheckReply = false;
-    SampleSetInitiator = 0;
-    for (byte i = 0; i < 4; ++i) {
-      IP[i] = 0;
-    }
-    ZERO_FILL(HostName);
-    ZERO_FILL(Publish);
-    ZERO_FILL(Subscribe);
-    ZERO_FILL(MQTTLwtTopic);
-    ZERO_FILL(LWTMessageConnect);
-    ZERO_FILL(LWTMessageDisconnect);
-  }
-
-  boolean       UseDNS;
-  byte          IP[4];
-  unsigned int  Port;
-  char          HostName[65];
-  char          Publish[129];
-  char          Subscribe[129];
-  char          MQTTLwtTopic[129];
-  char          LWTMessageConnect[129];
-  char          LWTMessageDisconnect[129];
-  unsigned int  MinimalTimeBetweenMessages;
-  unsigned int  MaxQueueDepth;
-  unsigned int  MaxRetry;
-  boolean       DeleteOldest; // Action to perform when buffer full, delete oldest, or ignore newest.
-  unsigned int  ClientTimeout;
-  boolean       MustCheckReply; // When set to false, a sent message is considered always successful.
-  uint8_t       SampleSetInitiator; // The first plugin to start a sample set.
-
-  void validate() {
-    if (Port > 65535) Port = 0;
-    if (MinimalTimeBetweenMessages < 1  ||  MinimalTimeBetweenMessages > CONTROLLER_DELAY_QUEUE_DELAY_MAX)
-      MinimalTimeBetweenMessages = CONTROLLER_DELAY_QUEUE_DELAY_DFLT;
-    if (MaxQueueDepth > CONTROLLER_DELAY_QUEUE_DEPTH_MAX) MaxQueueDepth = CONTROLLER_DELAY_QUEUE_DEPTH_DFLT;
-    if (MaxRetry > CONTROLLER_DELAY_QUEUE_RETRY_MAX) MaxRetry = CONTROLLER_DELAY_QUEUE_RETRY_MAX;
-    if (MaxQueueDepth == 0) MaxQueueDepth = CONTROLLER_DELAY_QUEUE_DEPTH_DFLT;
-    if (MaxRetry == 0) MaxRetry = CONTROLLER_DELAY_QUEUE_RETRY_DFLT;
-    if (ClientTimeout < 10 || ClientTimeout > CONTROLLER_CLIENTTIMEOUT_MAX) {
-      ClientTimeout = CONTROLLER_CLIENTTIMEOUT_DFLT;
-    }
-    ZERO_TERMINATE(HostName);
-    ZERO_TERMINATE(Publish);
-    ZERO_TERMINATE(Subscribe);
-    ZERO_TERMINATE(MQTTLwtTopic);
-    ZERO_TERMINATE(LWTMessageConnect);
-    ZERO_TERMINATE(LWTMessageDisconnect);
-  }
-
-  IPAddress getIP() const {
-    IPAddress host(IP[0], IP[1], IP[2], IP[3]);
-    return host;
-  }
-
-  String getHost() const {
-    if (UseDNS) {
-      return HostName;
-    }
-    return getIP().toString();
-  }
-
-  void setHostname(const String& controllerhostname) {
-    safe_strncpy(HostName, controllerhostname.c_str(), sizeof(HostName));
-    updateIPcache();
-  }
-
-  boolean checkHostReachable(bool quick) {
-    if (!WiFiConnected(10)) {
-      return false; // Not connected, so no use in wasting time to connect to a host.
-    }
-    delay(1); // Make sure the Watchdog will not trigger a reset.
-    if (quick && ipSet()) return true;
-    if (UseDNS) {
-      if (!updateIPcache()) {
-        return false;
-      }
-    }
-    return hostReachable(getIP());
-  }
-
-  boolean connectToHost(WiFiClient &client) {
-    if (!checkHostReachable(true)) {
-      return false; // Host not reachable
-    }
-    byte retry = 2;
-    bool connected = false;
-    while (retry > 0 && !connected) {
-      --retry;
-      // In case of domain name resolution error result can be negative.
-      // https://github.com/esp8266/Arduino/blob/18f643c7e2d6a0da9d26ff2b14c94e6536ab78c1/libraries/Ethernet/src/Dns.cpp#L44
-      // Thus must match the result with 1.
-      connected = connectClient(client, getIP(), Port);
-      if (connected) return true;
-      if (!checkHostReachable(false))
-        return false;
-    }
-    return false;
-  }
-
-  // Returns 1 if successful, 0 if there was a problem resolving the hostname or port
-  int beginPacket(WiFiUDP &client) {
-    if (!checkHostReachable(true)) {
-      return 0; // Host not reachable
-    }
-    byte retry = 2;
-    int connected = 0;
-    while (retry > 0 && connected == 0) {
-      --retry;
-      connected = client.beginPacket(getIP(), Port);
-      if (connected != 0) return connected;
-      if (!checkHostReachable(false))
-        return 0;
-      delay(10);
-    }
-    return 0;
-  }
-
-  String getHostPortString() const {
-    String result = getHost();
-    result += ":";
-    result += Port;
-    return result;
-  }
-
-private:
-  bool ipSet() {
-    for (byte i = 0; i < 4; ++i) {
-      if (IP[i] != 0) return true;
-    }
-    return false;
-  }
-
-  bool updateIPcache() {
-    if (!UseDNS) {
-      return true;
-    }
-    if (!WiFiConnected()) return false;
-    IPAddress tmpIP;
-    if (resolveHostByName(HostName, tmpIP)) {
-      for (byte x = 0; x < 4; x++) {
-        IP[x] = tmpIP[x];
-      }
-      return true;
-    }
-    return false;
-  }
-
-};
-
-typedef std::shared_ptr<ControllerSettingsStruct> ControllerSettingsStruct_ptr_type;
-#define MakeControllerSettings(T) ControllerSettingsStruct_ptr_type ControllerSettingsStruct_ptr(new ControllerSettingsStruct());\
-                                    ControllerSettingsStruct& T = *ControllerSettingsStruct_ptr;
 
 
 
@@ -1359,53 +673,6 @@ struct ExtraTaskSettingsStruct
   int16_t TaskDevicePluginConfig[PLUGIN_EXTRACONFIGVAR_MAX];
 } ExtraTaskSettings;
 
-/*********************************************************************************************\
- * EventStruct
-\*********************************************************************************************/
-struct EventStruct
-{
-  EventStruct() :
-    Data(NULL), idx(0), Par1(0), Par2(0), Par3(0), Par4(0), Par5(0),
-    Source(0), TaskIndex(TASKS_MAX), ControllerIndex(0), ProtocolIndex(0), NotificationIndex(0),
-    BaseVarIndex(0), sensorType(0), OriginTaskIndex(0) {}
-  EventStruct(const struct EventStruct& event):
-          String1(event.String1)
-        , String2(event.String2)
-        , String3(event.String3)
-        , String4(event.String4)
-        , String5(event.String5)
-        , Data(event.Data)
-        , idx(event.idx)
-        , Par1(event.Par1), Par2(event.Par2), Par3(event.Par3), Par4(event.Par4), Par5(event.Par5)
-        , Source(event.Source), TaskIndex(event.TaskIndex), ControllerIndex(event.ControllerIndex)
-        , ProtocolIndex(event.ProtocolIndex), NotificationIndex(event.NotificationIndex)
-        , BaseVarIndex(event.BaseVarIndex), sensorType(event.sensorType)
-        , OriginTaskIndex(event.OriginTaskIndex)
-         {}
-
-  String String1;
-  String String2;
-  String String3;
-  String String4;
-  String String5;
-  byte *Data;
-  int idx;
-  int Par1;
-  int Par2;
-  int Par3;
-  int Par4;
-  int Par5;
-  byte Source;
-  byte TaskIndex; // index position in TaskSettings array, 0-11
-  byte ControllerIndex; // index position in Settings.Controller, 0-3
-  byte ProtocolIndex; // index position in protocol array, depending on which controller plugins are loaded.
-  byte NotificationIndex; // index position in Settings.Notification, 0-3
-  //Edwin: Not needed, and wasnt used. We can determine the protocol index with getNotificationProtocolIndex(NotificationIndex)
-  // byte NotificationProtocolIndex; // index position in notification array, depending on which controller plugins are loaded.
-  byte BaseVarIndex;
-  byte sensorType;
-  byte OriginTaskIndex;
-};
 
 
 /*********************************************************************************************\
@@ -2222,82 +1489,9 @@ struct portStatusStruct {
 std::map<uint32_t, portStatusStruct> globalMapPortStatus;
 
 
-/********************************************************************************************\
-  Pre defined settings for off-the-shelf hardware
-  \*********************************************************************************************/
-
-// This enum will be stored, so do not change order or at least the values.
-enum DeviceModel {
-  DeviceModel_default = 0,
-  DeviceModel_Sonoff_Basic,
-  DeviceModel_Sonoff_TH1x,
-  DeviceModel_Sonoff_S2x,
-  DeviceModel_Sonoff_TouchT1,
-  DeviceModel_Sonoff_TouchT2,
-  DeviceModel_Sonoff_TouchT3,
-  DeviceModel_Sonoff_4ch,
-  DeviceModel_Sonoff_POW,
-  DeviceModel_Sonoff_POWr2,
-  DeviceModel_Shelly1,
-
-  DeviceModel_MAX
-};
-
-bool modelMatchingFlashSize(DeviceModel model, int size_MB);
 
 
-struct ResetFactoryDefaultPreference_struct {
-  ResetFactoryDefaultPreference_struct(uint32_t preference = 0) : _preference(preference) {}
 
-  DeviceModel getDeviceModel() const {
-    return static_cast<DeviceModel>(_preference & 0xFF);
-  }
-
-  void setDeviceModel(DeviceModel model) {
-    _preference &= ~(0xFF); // set DeviceModel bits to 0
-    _preference |= model;
-  }
-
-  bool keepWiFi() const { return getBitFromUL(_preference, 9); }
-  void keepWiFi(bool keep) {       setBitToUL(_preference, 9, keep); }
-
-  bool keepNTP() const { return getBitFromUL(_preference, 10); }
-  void keepNTP(bool keep) {       setBitToUL(_preference, 10, keep); }
-
-  bool keepNetwork() const { return getBitFromUL(_preference, 11); }
-  void keepNetwork(bool keep) {       setBitToUL(_preference, 11, keep); }
-
-  bool keepLogSettings() const { return getBitFromUL(_preference, 12); }
-  void keepLogSettings(bool keep) {       setBitToUL(_preference, 12, keep); }
-
-  bool keepUnitName() const { return getBitFromUL(_preference, 13); }
-  void keepUnitName(bool keep) {       setBitToUL(_preference, 13, keep); }
-
-  // filenr = 0...3 for files rules1.txt ... rules4.txt
-  bool fetchRulesTXT(int filenr) const { return getBitFromUL(_preference, 14 + filenr); }
-  void fetchRulesTXT(int filenr, bool fetch) {       setBitToUL(_preference, 14 + filenr, fetch); }
-
-  bool fetchNotificationDat() const { return getBitFromUL(_preference, 18); }
-  void fetchNotificationDat(bool fetch) {       setBitToUL(_preference, 18, fetch); }
-
-  bool fetchSecurityDat() const { return getBitFromUL(_preference, 19); }
-  void fetchSecurityDat(bool fetch) {       setBitToUL(_preference, 19, fetch); }
-
-  bool fetchConfigDat() const { return getBitFromUL(_preference, 20); }
-  void fetchConfigDat(bool fetch) {       setBitToUL(_preference, 20, fetch); }
-
-  bool deleteFirst() const { return getBitFromUL(_preference, 21); }
-  void deleteFirst(bool checked) {       setBitToUL(_preference, 21, checked); }
-
-  
-
-  uint32_t getPreference() { return _preference; }
-
-  // TODO TD-er: Add extra flags for settings to keep/set when reset to default.
-
-private:
-  uint32_t _preference;
-} ResetFactoryDefaultPreference;
 
 void applyFactoryDefaultPref() {
   // TODO TD-er: Store it in more places to make it more persistent
@@ -2557,10 +1751,6 @@ static String LoRa_addFloat(float value, PackedData_enum datatype) {
   return LoRa_base16Encode(data, cursor);
 }
 
-
-// Forward declarations PackedData related functions
-String getPackedFromPlugin(struct EventStruct *event, uint8_t sampleSetCount);
-
 #endif // USES_PACKED_RAW_DATA
 
 
@@ -2571,6 +1761,29 @@ String getPackedFromPlugin(struct EventStruct *event, uint8_t sampleSetCount);
 #define SPIFFS_CHECK(result, fname) if (!(result)) { return(FileError(__LINE__, fname)); }
 #include "WebServer_Rules.h"
 
+#ifdef USES_BLYNK
+// Blynk_get prototype
+//boolean Blynk_get(const String& command, byte controllerIndex,float *data = NULL );
+
+int firstEnabledBlynkController() {
+  for (byte i = 0; i < CONTROLLER_MAX; ++i) {
+    byte ProtocolIndex = getProtocolIndex(Settings.Protocol[i]);
+    if (Protocol[ProtocolIndex].Number == 12 && Settings.ControllerEnabled[i]) {
+      return i;
+    }
+  }
+  return -1;
+}
+#endif
+
+// These have to be at the end of this .h file for now, 
+// since they use functions or objects otherwise not yet declared
+
+
+// Must be included after all the defines, since it is using TASKS_MAX
+#include "_Plugin_Helper.h"
+// Plugin helper needs the defined controller sets, thus include after 'define_plugin_sets.h'
+#include "_CPlugin_Helper.h"
 
 
 #endif /* ESPEASY_GLOBALS_H_ */
