+++ src/ESPEasy.ino
@@ -84,25 +84,20 @@
 
 // Define globals before plugin sets to allow a personal override of the selected plugins
 #include "ESPEasy-Globals.h"
-// Must be included after all the defines, since it is using TASKS_MAX
-#include "_Plugin_Helper.h"
-// Plugin helper needs the defined controller sets, thus include after 'define_plugin_sets.h'
-#include "_CPlugin_Helper.h"
 
+
+#ifdef USES_BLYNK
 // Blynk_get prototype
 boolean Blynk_get(const String& command, byte controllerIndex,float *data = NULL );
 
-int firstEnabledBlynkController() {
-  for (byte i = 0; i < CONTROLLER_MAX; ++i) {
-    byte ProtocolIndex = getProtocolIndex(Settings.Protocol[i]);
-    if (Protocol[ProtocolIndex].Number == 12 && Settings.ControllerEnabled[i]) {
-      return i;
-    }
-  }
-  return -1;
-}
+int firstEnabledBlynkController();
+#endif
 
 //void checkRAM( const __FlashStringHelper* flashString);
+
+#ifdef CORE_POST_2_5_0
+void preinit();
+#endif
 
 #ifdef CORE_POST_2_5_0
 /*********************************************************************************************\
@@ -384,6 +379,13 @@ void setup()
 }
 
 #ifdef USE_RTOS_MULTITASKING
+void RTOS_TaskServers( void * parameter );
+void RTOS_TaskSerial( void * parameter );
+void RTOS_Task10ps( void * parameter );
+void RTOS_HandleSchedule( void * parameter );
+#endif
+
+#ifdef USE_RTOS_MULTITASKING
 void RTOS_TaskServers( void * parameter )
 {
  while (true){
@@ -582,29 +584,8 @@ void flushAndDisconnectAllClients() {
   process_serialWriteBuffer();
 }
 
+
 #ifdef USES_MQTT
-void runPeriodicalMQTT() {
-  // MQTT_KEEPALIVE = 15 seconds.
-  if (!WiFiConnected(10)) {
-    updateMQTTclient_connected();
-    return;
-  }
-  //dont do this in backgroundtasks(), otherwise causes crashes. (https://github.com/letscontrolit/ESPEasy/issues/683)
-  int enabledMqttController = firstEnabledMQTTController();
-  if (enabledMqttController >= 0) {
-    if (!MQTTclient.loop()) {
-      updateMQTTclient_connected();
-      if (MQTTCheck(enabledMqttController)) {
-        updateMQTTclient_connected();
-      }
-    }
-  } else {
-    if (MQTTclient.connected()) {
-      MQTTclient.disconnect();
-      updateMQTTclient_connected();
-    }
-  }
-}
 
 void updateMQTTclient_connected() {
   if (MQTTclient_connected != MQTTclient.connected()) {
@@ -634,6 +615,29 @@ void updateMQTTclient_connected() {
   setIntervalTimer(TIMER_MQTT);
 }
 
+void runPeriodicalMQTT() {
+  // MQTT_KEEPALIVE = 15 seconds.
+  if (!WiFiConnected(10)) {
+    updateMQTTclient_connected();
+    return;
+  }
+  //dont do this in backgroundtasks(), otherwise causes crashes. (https://github.com/letscontrolit/ESPEasy/issues/683)
+  int enabledMqttController = firstEnabledMQTTController();
+  if (enabledMqttController >= 0) {
+    if (!MQTTclient.loop()) {
+      updateMQTTclient_connected();
+      if (MQTTCheck(enabledMqttController)) {
+        updateMQTTclient_connected();
+      }
+    }
+  } else {
+    if (MQTTclient.connected()) {
+      MQTTclient.disconnect();
+      updateMQTTclient_connected();
+    }
+  }
+}
+
 int firstEnabledMQTTController() {
   for (byte i = 0; i < CONTROLLER_MAX; ++i) {
     byte ProtocolIndex = getProtocolIndex(Settings.Protocol[i]);
@@ -643,8 +647,8 @@ int firstEnabledMQTTController() {
   }
   return -1;
 }
-#endif //USES_MQTT
 
+#endif //USES_MQTT
 
 /*********************************************************************************************\
  * Tasks that run 50 times per second
